requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
boto3==1.42.91
botocore==1.42.91
certifi==2026.2.25
click==8.3.2
fastapi==0.136.0
h11==0.16.0
idna==3.11
jmespath==1.1.0
massive==2.5.0
numpy==2.4.4
pandas==3.0.2
psycopg2-binary==2.9.11
pyarrow==23.0.1
pydantic==2.13.2
pydantic_core==2.46.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.44.0
websockets==16.0
